Update patch set 3
Patch Set 3:

Hi Jorge
This appears to be a timing issue. It works 100% of the time in my local environment but fails intermittently on Jenkins. Adding a 5 second sleep appears to fix the issue.
Can you add:
sleep 5
after line 29 in plans/policy/apex-pdp/setup.sh
and also add "#" to line24 as this line is intended to be a comment, i.e.

24 #Wait for initialization
25 for i in {1..10}; do
26    curl -sS ${APEX_IP}:23324 && break
27    echo sleep $i
28    sleep $i
29 done
30 sleep 5

Patch-set: 3