Update patch set 1
Patch Set 1:

As a note Ram, for the other images we produce a "major"."minor" docker image.   I thought that was a recommendation.   I noticed that it does not get generated for the policy-distribution image.   We should probably add it to our backlog.  It is convenient since don't have to worry of the z in the x.y.z bumps.

Patch-set: 1
Label: Verified=0