Update patch set 1
Patch Set 1:

If the only thing that holds this up is https then I don't see any reason we can't merge this so that the test start passing. I can come back and modify it with https once I get that to work in xacml pdp, which I am working on now.  The other option is that I modify the ci-management jenkins job list to remove the csit tests from the job list.

Patch-set: 1