Update patch set 1
Patch Set 1:

Hi Bilal, this is great.   Let's try to do some clean up to have a clean base for all the tests.   For these ones, we don't policy/engine related ones, meaning: brmsgw, pap, pdp.

The mock-service in the set up, not sure if it's necessary but if not, can be deleted.

The only components that we will play with are : mariadb, nexus, and drools-pdp.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0