Update patch set 2
Patch Set 2:

I think for the *drools-pdp* the changes are good, we can leave the same model as it was doing initially for now.

For the *drools-application*, I really like the way that Bilal has structured the work with the docker-componse in the setup.   I think it is better to keep the same structure as in the other projects.

Patch-set: 2
Label: Code-Review=0