Update patch set 5
Patch Set 5:

> I think this should be cherry picked to dublin also. Bilal, please
 > keep track of this one also.

I don't know if this can be cherry picked or not, as it removes version numbers used when pulling the docker images.  Does the dublin csit know how to pick the dublin docker images?  If not, then the dublin versions will have to be added to all of the policy yml files in the dublin csit (not just the ones included in this review).  On the other hand, if we don't care that the dublin csit picks dublin docker images, then we should be able to cherry pick this.

This also requires changes to a number of the policy repos - reviews can be found in POLICY-1829.  (They all have titles of the form, "Fix XXX docker chmod.)  I think those can all be cherry picked.

Patch-set: 5