Pull correct docker images for master branch
Enables Policy CSITs for API, PAP, Distribution, Drools-applications,
Apex-PDP and XACML-PDP to pull correct docker images based on the
code branch.

--------------

Updating the policy model json in apex-pdp csit

Changes were made to read the policy model as proper json instead of
json formatted string in APEX-PDP. With this fix in place, the policy
tosca json file need to be updated with the proper json.

Issue-ID: POLICY-1836
Issue-ID: POLICY-1796
Change-Id: I6b99c8ae9cb78dd8ebddc522bc2a7080ec6e5735
Signed-off-by: a.sreekumar <ajith.sreekumar@est.tech>
Signed-off-by: Bilal A <bilal@research.att.com>