Update patch set 4
Patch Set 4:

(3 comments)

Now that tosca-lab container is part of the test env, generation of Tosca files could actually be added in the next step. As for the documentation - some of these parts are too much in the air for me to say confidently anything about them yet. This partition (especially regarding some details of the env setup) between core and DCAE-DS is a suggestion for which I'll need confirmation from Ofir Sonsino, Avi Ziv etc. so I'd rather have that sorted out before writing it down... but, then again, that's why this is currently in "work in progress" state

Patch-set: 4
Label: Code-Review=0