Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Isn't this subsumed by Bilal's updates in the following review?
> 
>   https://gerrit.onap.org/r/#/c/integration/csit/+/90602/

This is for dublin branch. I saw a review from Bilal for all the other repos except apex which was merged a few days ago. For Apex, currently, docker image with latest tag is being pulled in csit tests.

Also,unlike other repos, Apex docker image tagging was bit different in dublin(not anymore). So, in this change, that is also anticipated in the setup.sh script.

Patch-set: 2
Label: Code-Review=0