Update patch set 5
Patch Set 5: Code-Review-1

One problem is that the jinja templater expects values of the form:
  {{ name }}
while the old templater expects values of the form:
  ${name}
Consequently, all of the template files must be modified.  After making that change in my local copy, it then told me that "Apply Template" expects 3 arguments, but only 2 were given.  In looking at TemplatingEngine, it appears that the object must be constructed by passing a template location argument.  Then perhaps the newly constructed object must be passed to the "Apply Template" as its first argument?

Patch-set: 5