Update patch set 2
Patch Set 2:

I made these corrections to complete the PORTAL-702 ticket in the wake of https://gerrit.onap.org/r/#/c/testsuite/+/93213/, but the more I looked into this the more confused I got. The tests in test1.robot are not used except for healtheck, the keywords are duplicated in portal_VID.robot (which is not used from anywhere) and those keywords are riddled with so many fatal typos that they wouldn't even work if called. Why are all these files and keywords here to begin with?

Patch-set: 2