Add support for https tests for SDC-DCAE-D
This commit introduces a new set of variables for sdc-dcae-d test in the
same style as was for sdc tests. Some refactor had to be done to make it
work and make it less copy-pasted.

BEWARE: Although it was successfully tested it depends on a commit which
is in another repo and was not merged at the time of this writing:
https://gerrit.onap.org/r/#/c/sdc/+/96062/

BEWARE 2: The robotframework-onap pkg installed by pip needs this fix:
https://gerrit.onap.org/r/#/c/testsuite/python-testing-utils/+/94306/

Which was merged and backported to version '0.5' but it was not applied
to the actual package in nexus3.onap.org:
https://nexus3.onap.org/repository/PyPi.staging/packages/
robotframework-onap/0.5/robotframework_onap-0.5-py2.py3-none-any.whl

Issue-ID: SDC-2564

Change-Id: Ic0725dbc4643d0563594f28d2a772968e291f788
Signed-off-by: Petr Ospalý <p.ospaly@partner.samsung.com>