Fix SDC_LOCAL_GITREPO for relative paths
SDC_LOCAL_GITREPO will not do what you want (most likely) if used
with a relative path - it worked only with an absolute one.

This commit fixes this but it adds a new requirement: 'realpath';
which should be on any system as it is part of the coreutils and
similar package sets.

Issue-ID: SDC-2553

Signed-off-by: Petr Ospalý <p.ospaly@partner.samsung.com>
Change-Id: I2d79252cb315538997c91b754acb59f52469df00