Update patch set 5
Patch Set 5:

Once Liam's change to return 406 is in for API, two of the Create tests in API should be passing. 2 Delete tests will still be failing I assume because of the bug I mentioned in RocketChat. Either we have to fix that bug, or change the order of delete (delete v2 first and then v1)

Patch-set: 5
Label: Code-Review=0