Update patch set 1
Patch Set 1:

The root cause for the failing nightly job is very unclear as the sudden "Build step 'Execute shell' marked build as failure" comes from nowhere - there is nothing in the entire log to explain what causes it. The matter is even more mysterious given the fact that "nightly" is practically identical with "sanity", which passed just fine (I don't even know what we have the nightly for). I'll just recheck this first to see if this repeats.

Patch-set: 1