Use wait-for-dependencies in drools-apps CSIT
Revised drools-apps yml to conform the pattern used for api, pap,
and the other policy repos.  In particular, deleted the port check
from the setup script and replaced it with docker wait-for-dependency.
With this approach, the docker containers are brought up in stages
so that dependencies are met (i.e., doesn't bring up drools until
the mariadb server is listening on its port).  The net result is
that the CSIT runs much faster.

Issue-ID: POLICY-2434
Change-Id: I7be2c7e7595cdc1541729a9675255b379565ff2e
Signed-off-by: Jim Hahn <jrh3@att.com>