Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> Looks good! A couple questions, why the api and pap components have the overriding keystore and truststore?  and logback ?

Is there a way to just overlay a single file?  I tried it, but it complained that it was overlaying a file that already existed.  As a result, I resorted to overlaying the whole directory, which requires the other files to be included within the overriding directory.

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>