Update patch set 1
Patch Set 1:

This is going to need some more refactoring of the "so" directory name.
The JJB regex lookup for "so" picks up this change and runs the SO jobs as well.
We had a similar issue before and changed the dir name to "sorch" as the so jobs kept failing, blocking the SWUP changes.

Patch-set: 1
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>