Update patch set 9
Patch Set 9:

(1 comment)

> Patch Set 9:
> 
> (1 comment)
> 
> > Patch Set 9: Code-Review+1
> > 
> > As a bonus, now we also see the CSIT jobs under https://jenkins.onap.org/view/ccsdk-oran/ as well
> 
> Please take a look at rsync call in that job:
> 
>  rsync -av / /w/workspace/ccsdk-oran-master-csit-polmansuite/archives
> 
> Are directories set properly? Did you intend to sync whole root?

I guess this comment was for Lasse K, not sure what it means since it seems to be related to the CSIT environment. But it is true that a repo is cloned into the archives dir during test execution (setup script) and the intention is to save this cloned repo for trouble shooting when test is complete, some important logs are created in the directory structure in that cloned repo.

Patch-set: 9