Make stand-alone Policy environment
This is a first cut at making it possible to easily run Policy
components within a stand-alone environment.  Made the following
changes:
- enhanced wait_for_port.sh to accept a list of hosts/ports
- enhanced wait_for_port.sh to accept an optional command to execute
  when the wait completes
- removed "extra" services from the drools-apps yml so that the entire
  thing can be brought up in a single invocation
- modified the drools-apps setup.sh

With these changes, a developer can clone the csit repo, modify
the docker image versions in the yml file, and then bring up the
components using:

  docker-compose -f scripts/policy/drools-apps/docker-compose-drools-apps.yml
    up --detach

Note: this does not preload or deploy any policies; that must be done
manually via curl.

Issue-ID: POLICY-2742
Change-Id: I98b1fb2fcc00c2cc5edbdc3ecf87cb213673fb18
Signed-off-by: Jim Hahn <jrh3@att.com>