Update patch set 7
Patch Set 7: Code-Review+1

I'm marking this for review.  However, if folks don't feel like this is a good approach, then we can reject it and take a different approach when I return.  For minor issues, feel free to patch it and merge it - no need to wait for me to get back.

With this, the S3P tests can be modified to use the CSIT, using drools-compose to bring up the relevant component(s).  As for moving the CSITs into the individual repos, I'm still a bit unsure about that, as there is a lot of common code between the CSITs for each component.  We'll have to come up with a good way to share code (perhaps leave the common code in integration/csit?) between the repos.  I also don't want CSIT failures to prevent jenkins verification or merge jobs from succeeding, otherwise we might wind up in a catch-22 situation.

Patch-set: 7
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1
Work-in-progress: false