Update patch set 7
Patch Set 7:

Nice work!
I am only a bit concerned about the simulators, in most cases we only need message-router. Here since all are in one container, there is the overhead to run everything, also we can't name the service as message-router, and hence making it necessary to mount the defaultConfig files for all components.
But maybe its a tradeoff. This definitely make the scripts clean and easy to use.

Patch-set: 7