Update patch set 7
Patch Set 7:

(1 comment)

For a standalone policy environment, I think it would be better if possible to have it in the policy/docker repo as what we had in earlier releases although unmaintained.   Having it in the integration/csit, besides not having full control (we don't own the repo), poses the additional overhead of depending on the csit software framework, which again we don't have control over it, nor it would be necessary for a standalone docker-compose based environment.   A thought to consider.

Patch-set: 7