Update patch set 7
Patch Set 7:

(2 comments)

> Patch Set 7:
> 
> (1 comment)
> 
> For a standalone policy environment, I think it would be better if possible to have it in the policy/docker repo as what we had in earlier releases although unmaintained.   Having it in the integration/csit, besides not having full control (we don't own the repo), poses the additional overhead of depending on the csit software framework, which again we don't have control over it, nor it would be necessary for a standalone docker-compose based environment.   A thought to consider.

Figured we'd move things around in the next round of reviews for this.  Is policy/docker an appropriate place for this (seems kind of like creating a circular dependency, but probably ok in this case)?  As for our CSITs, it makes sense to me that we'd distribute them among the various repos, but I don't want a CSIT failure to prevent a verify or merge job from completing successfully or we could wind up in a "catch-22" situation.  Not sure if we can do that easily, or if we should just set up a policy/csit repo.

Patch-set: 7
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>