Update patch set 2
Patch Set 2:

Yeah, I'm having the same issue with replies, but want to make sure it isn't merged, at least for the moment.

Honestly, I don't know that this will actually fix the problem.  For that matter, it might even break some of the component CSITs.  All I know is that our CSITs are failing due to the rate limit when trying to download sonatype/nexus, and this change appears to make it download from nexus3 (not even sure why it downloads sonatype/nexus).  As for the onap images, our docker compose yaml already specifies nexus3, so that shouldn't be rate limited anyway.  I'm wondering if it's just all of the jobs combined, not just CSITs, but the docker builds, etc., that use dockerhub for sonatype/nexus and alpine base images.  If that's the case, then this may fix the CSITs while leaving those other jobs broken.

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>