Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > What CSIT failures is this title referring to?
> 
> Hi Lasse, when pulling down Master CSIT tests and running the "integration-etsi-testing" we are seeing failures on our local system due to the NFVO NS LCM Container being unable to populate the MariaDB NFVO Tables.

I can email you logs of the errors if you would like, it appears that the Flyway Dependency we use to populate the MariaDB Tables no longer supports MariaDB version 10.1 or older, shown by this error which appears in the SO-ETSI-NFVO-NS-LCM Container logs:

"org.flywaydb.core.internal.license.FlywayEditionUpgradeRequiredException: Flyway Pro Edition or MariaDB upgrade required: MariaDB 10.1 is no longer supported by Flyway Community Edition, but still supported by Flyway Pro Edition."

Upgrading the MariaDB Version then caused additional errors with the table creation for the NFVO DB within MariaDB. This was due to the Camunda Migration from 7.9 to 7.10, so in order to fix this we needed to edit the mariadb_engine_sql script, which creates the Camunda Database and tables. Happy to expand on this over email if easier.

Patch-set: 1