Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> BTW,
> 
> > > The Verify build appears to have failed the so-master-verify-csit-integration-etsi-testing job (https://jenkins.onap.org/job/so-master-verify-csit-integration-etsi-testing/191/)
> 
> Is this failure the same you were getting locally?

Yes this is the same failure we were getting locally. It continues to wait for the pods to spin up but eventually times out. Then if we simply updated the MariaDB version to one greater than 10.1.11 we then received errors from on the NFVO container when it attempted to create it's tables in the MariaDB

Patch-set: 3