Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > I think the python versioning issue needs to be addressed.
> > Tried to join the integration meeting yesterday at 2pm UTC but could join. Has the time been changed?
> 
> https://wiki.onap.org/display/DW/Integration+Project says UTC 13:00 (this has remained the same for me for months, but not sure what daylight savings changes might have done in other time zones)
> 
> Python versioning was discussed and there's also some related correction done to DMaaP DataRouter CSIT by Alexander Mazuruk here: https://gerrit.onap.org/r/c/integration/csit/+/118753. The version used in Jenkins should be clarified with LF (@Morgan, should we/did we raise any query ticket about it yet?) but I'm not sure that topic should be subject of this commit?

Could be an issue with my system but when I add the ics file it adds it for 2:00pm for some reason.
How does that patch relate to the python version used?
I guess the fact that the jobs are passing running py2 is a false positive as the mocks aren't coming up. We need to try get it running on py3 ideally.

Patch-set: 4
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>